+++ packages/zod/src/index.ts
@@ -1,4 +1,5 @@
 export * from './global-realm';
+export * from './re-exports';
 export * from './realm';
 export * from './ref';
 export * from './types';